DEFw: Initial Commit
DEFw - Distributed Execution Framework
This Framework is build upon python and allows users to build services.
These services expose an API which clients can call in order to use the
services.
It implements an approach where python classes are used to describe the
services which can be remotely executed.
The python interpreter is integrated in a C infrastructure which
provides the lower level communication between the different instances
of the defw.
Only sockets are supported at the moment, but support for libfabric is
planned such that this infrastructure can operate on any libfabric
supported NIC.

Signed-off-by: Amir Shehata <amir.shehata@gmail.com>