DEFw: Modify the remote execution handling
Originally the DEFw remote execution took the endpoint as input.
However, this forced the base remote class to make assumption on the
naming convention to find the service.

This has been modified to take in the service info which we need to
connect to. The service info contains the name of the service module
which we'll need to connect to as well as the endpoint.

This allows the remote base class to connect to the endpoint as well as
have the proper module name to request to send to the remote in order
for the remote to import that module and instantiate the class which is
required.

Signed-off-by: Amir Shehata <amir.shehata@gmail.com>