DEFw: Update execution and exit modes
Have 2 execution modes. If the program is run with the binary name
"python" or "pythonX", behave exactly as a python interpreter. This
allows applications to use DEFw as a python interpreter and get all the
other features with it.

If the program is run as defwp (or anything else) then parse the command
lines and process them as you would normally.

Convert all the python threads in daemons to avoid having the
interpreter wait on them until they exit. In python mode this causes a
problem, because there is no way to call the me.exit() function which
shutsdown all internal threads.

Signed-off-by: Amir Shehata <amir.shehata@gmail.com>